requirements.txt
asgiref==3.6.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
django-import-export==3.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
gunicorn==20.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
psycopg2==2.9.5
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.3
tablib==3.3.0
tzdata==2022.7
whitenoise==6.4.0
xlrd==2.0.1
xlwt==1.3.0
